% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipage.R
\name{layout_facets}
\alias{layout_facets}
\title{Create multipage layout from facets}
\usage{
layout_facets(gg, facet_var, ncol = NULL, nrow = NULL, width = NULL,
  height = NULL, ...)
}
\arguments{
\item{gg}{A ggplot}

\item{facet_var}{Descr.}

\item{ncol}{Descr.}

\item{nrow}{Descr.}

\item{width}{Descr.}

\item{height}{Descr.}

\item{...}{Descr.}
}
\description{
Create multipage layout from facets
}
