% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipage.R
\name{save_multipage}
\alias{save_multipage}
\title{Save multipage layout to file}
\usage{
save_multipage(gg = last_plot(), filename, device = NULL,
  path = NULL, scale = 1, width = NA, height = NA,
  units = c("mm", "cm", "in"), dpi = 300, limitsize = TRUE,
  return_input = FALSE, multiple_files = FALSE, ...)
}
\arguments{
\item{gg}{ggplot or list of ggplots}

\item{filename}{File name to create on disk.}

\item{device}{Device to use. Can either be a device function
(e.g. \code{\link[=png]{png()}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{scale}{Multiplicative scaling factor.}

\item{width}{desr}

\item{height}{desr}

\item{units}{desr}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{return_input}{Return the input ggplot or plotlist is after saving.
This enables the use within \code{dplyr} pipes.}

\item{multiple_files}{Save pages as individal files.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\description{
This function takes a ggplot (for single page) or list of ggplots (for multi page) and writes them to file.
In case the input has absolute dimensions, width and height of the output device are adjusted to fit the content.
}
